+++ app/src/main/java/com/kova700/zerotomvvm/PokemonListItemDiffUtil.kt
@@ -1,0 +1,13 @@
+package com.kova700.zerotomvvm
+
+import androidx.recyclerview.widget.DiffUtil
+
+class PokemonListItemDiffUtil : DiffUtil.ItemCallback<PokemonListItem>() {
+    override fun areItemsTheSame(oldItem: PokemonListItem, newItem: PokemonListItem): Boolean {
+        return oldItem.pokemon.name == newItem.pokemon.name
+    }
+
+    override fun areContentsTheSame(oldItem: PokemonListItem, newItem: PokemonListItem): Boolean {
+        return oldItem == newItem
+    }
+}
